+++ tests/assertions/assert_inner_html.test.js
@@ -47,14 +47,14 @@ describe("Assert Inner Html", function() {
         await browser.open(configUrls.index);
         await utils.assertThrowsAssertionAsync (async () => {
             await browser.assert.innerHtml("b", "not-text");
-        }, `Expected element "b" to have inner html "not-text", "Hidden text " found.`);
+        }, `Expected element "b" to have inner html "not-text", "Hidden text " found.`, "Hidden text,", "not-text");
     });
 
     it("InnerHtml Throws With Custom Message", async () => {
         await browser.open(configUrls.index);
         await utils.assertThrowsAssertionAsync (async () => {
             await browser.assert.innerHtml("b", "not-text", "html fails");
-        }, "html fails");
+        }, "html fails", "Hidden text,", "not-text");
     });
 
     it("InnerHtml Throws Element Not Found", async () => {